requirements.txt
aiofiles==24.1.0
airtest==1.3.5
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.8.0
apple-compress==0.2.3
arrow==1.3.0
asn1==2.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blessed==1.20.0
bpylist2==4.1.1
cached-property==1.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colored==2.3.0
coloredlogs==15.0.1
construct==2.10.70
cryptography==44.0.1
daemonize==2.5.0
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
developer_disk_image==0.0.4
dnspython==2.7.0
editor==1.6.6
email_validator==2.2.0
enum-compat==0.0.3
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
facebook-wda==1.5.0
fastapi==0.115.9
fastapi-cli==0.0.7
ffmpeg-python==0.2.0
filelock==3.17.0
future==1.0.0
gpxpy==1.5.0
h11==0.14.0
hexdump==3.3
hrpc==1.0.9
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
IceSpringPySideStubs-PySide6==1.3.1
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
iniconfig==2.0.0
inquirer3==0.6.1
ipsw-parser==1.4.0
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
la-panic==0.5.0
loguru==0.7.3
logzero==1.7.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mss==6.1.0
nest-asyncio==1.6.0
numpy==1.26.4
opack2==0.0.1
opencv-contrib-python==4.6.0.66
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parameter-decorators==0.0.2
parso==0.8.4
perfcat==1.0.10
pexpect==4.9.0
pillow==10.4.0
pluggy==1.5.0
plumbum==1.9.0
pocoui==1.0.94
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-python-adb==0.3.0.dev0
pure_eval==0.2.3
py==1.11.0
pycparser==2.22
pycrashreport==1.2.6
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pygnuutils==0.1.1
pyimg4==0.8.7
pykdebugparser==1.2.4
pylzss==0.3.4
pymobiledevice3==4.20.14
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
pytest==8.3.4
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pcapng==2.1.1
python-xlib==0.33
pytun-pmd3==2.0.10
pytz==2025.2
pyusb==1.3.1
pywinauto==0.6.3
PyYAML==6.0.2
qh3==1.4.1
readchar==4.2.1
remotezip2==0.0.2
requests==2.32.3
retry==0.9.2
rich==13.9.4
rich-toolkit==0.13.2
runs==1.2.2
shellingham==1.5.4
shiboken6==6.8.2.1
simple-tornado==0.2.2
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
srptools==1.0.1
sslpsk-pmd3==1.0.3
stack-data==0.6.3
starlette==0.45.3
tabulate==0.9.0
termcolor==2.5.0
tidevice==0.12.10
tidevice3==0.11.3
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
WebOb==1.8.9
websocket-client==0.48.0
websockets==15.0
wrapt==1.17.2
wsproto==1.2.0
xmod==1.8.1
xonsh==0.19.2
zeroconf==0.132.2
zipp==3.21.0
